+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,11 @@
         android:theme="@style/Theme.Contacts"
         tools:targetApi="31">
         <activity
+            android:name=".ContactAdd"
+            android:exported="false">
+
+        </activity>
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
