+++ app/src/main/res/values/strings.xml
@@ -11,4 +11,5 @@
     <string name="memo">메모</string>
     <string name="cancel">취소</string>
     <string name="save">저장</string>
+    <string name="contactGuide">오른쪽 + 버튼을 클릭하여\n 연락처를 등록해보세요.</string>
 </resources>
